FullyRefactored diary card widget
