+++ lib/ui/diary_home_page/diary_home_page_view.dart
@@ -3,7 +3,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:training_app/ui/diary_home_page/diary_home_page_bloc.dart';
 import 'package:training_app/ui/diary_home_page/diary_home_page_event.dart';
 import 'package:training_app/ui/diary_home_page/diary_home_page_state.dart';
-import 'package:training_app/ui/widget/diary_card/diary_card_provider.dart';
+import 'package:training_app/ui/widget/diary_card/diary_card.dart';
+// import 'package:training_app/ui/widget/diary_card/diary_card_provider.dart';
+// import 'package:training_app/ui/widget/diary_card/diary_card_view.dart';
 
 class DiaryHomeScreenView extends StatelessWidget {
   final String name;
@@ -168,13 +170,22 @@ class DiaryHomeScreenView extends StatelessWidget {
                   BlocBuilder<DiaryHomeScreenBloc, DiaryHomeScreenState>(
                     builder: (context, state) {
                       return Column(
-                        children: state.entries.map((entry) {
-                            return DiaryCardProvider(
-                            title: entry.title,
-                            username: entry.username,
-                            description: entry.description,
-                          );
-                        }).toList(),
+                        // children: state.entries.map((entry) {
+                        //     return DiaryCardProvider(
+                        //     title: entry.title,
+                        //     username: entry.username,
+                        //     description: entry.description,
+                        //   );
+                        // }).toList(),
+
+                    children: state.entries.map((entry) {
+                    return DiaryCard(
+                      title: entry.title,
+                      // username: 'John Doe',
+                      username: entry.username,
+                      description: entry.description,
+                    );
+                  }).toList(),
                       );
                     },
                   ),
