--- lib/ui/widget/diary_card/diary_card_event.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-@immutable
-abstract class DiaryCardEvent {}
-
-class ToggleShowMore extends DiaryCardEvent {
-  final bool showMore;
-
-  ToggleShowMore({
-    required this.showMore,
-  });
-}
