--- lib/ui/widget/diary_card/diary_card_state.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-@immutable
-class DiaryCardState {
-  final bool showMore;
-
-  const DiaryCardState({
-    required this.showMore,
-  });
-
-  static DiaryCardState get initialState => const DiaryCardState(
-        showMore: false,
-      );
-
-  DiaryCardState clone({
-    bool? showMore,
-  }) {
-    return DiaryCardState(
-      showMore: showMore ?? this.showMore,
-    );
-  }
-}
